chore: streamline TensorRT conversion scripts
